adjust folder structure (#34)
