+++ sdk/feature-management/.eslintrc
