+++ sdk/feature-management/package.json
