+++ sdk/feature-management/src/common/ITargetingContext.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export interface ITargetingContext {
+    userId?: string;
+    groups?: string[];
+}
+
