+++ sdk/feature-management/src/variant/Variant.ts
@@ -1,0 +1,9 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT license.
+
+export class Variant {
+    constructor(
+        public name: string,
+        public configuration: unknown
+    ) {}
+}
