+++ sdk/feature-management/test/sampleFeatureFlags.ts
@@ -1,0 +1,532 @@
+export enum Features {
+    VariantFeatureDefaultDisabled = "VariantFeatureDefaultDisabled",
+    VariantFeatureDefaultEnabled = "VariantFeatureDefaultEnabled",
+    VariantFeaturePercentileOn = "VariantFeaturePercentileOn",
+    VariantFeaturePercentileOff = "VariantFeaturePercentileOff",
+    VariantFeatureAlwaysOff = "VariantFeatureAlwaysOff",
+    VariantFeatureUser = "VariantFeatureUser",
+    VariantFeatureGroup = "VariantFeatureGroup",
+    VariantFeatureNoVariants = "VariantFeatureNoVariants",
+    VariantFeatureNoAllocation = "VariantFeatureNoAllocation",
+    VariantFeatureAlwaysOffNoAllocation = "VariantFeatureAlwaysOffNoAllocation",
+    VariantFeatureBothConfigurations = "VariantFeatureBothConfigurations",
+    VariantFeatureInvalidStatusOverride = "VariantFeatureInvalidStatusOverride",
+    VariantFeatureInvalidFromTo = "VariantFeatureInvalidFromTo",
+    VariantImplementationFeature = "VariantImplementationFeature",
+}
+
+export const featureFlagsConfigurationObject = {
+    "feature_management": {
+        "feature_flags": [
+            {
+                "id": "OnTestFeature",
+                "enabled": true
+            },
+            {
+                "id": "OffTestFeature",
+                "enabled": false
+            },
+            {
+                "id": "ConditionalFeature",
+                "enabled": true,
+                "conditions": {
+                    "client_filters": [
+                        {
+                            "name": "Test",
+                            "parameters": {
+                                "P1": "V1"
+                            }
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "ContextualFeature",
+                "enabled": true,
+                "conditions": {
+                    "client_filters": [
+                        {
+                            "name": "ContextualTest",
+                            "parameters": {
+                                "AllowedAccounts": [
+                                    "abc"
+                                ]
+                            }
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "AnyFilterFeature",
+                "enabled": true,
+                "conditions": {
+                    "requirement_type": "Any",
+                    "client_filters": [
+                        {
+                            "name": "Test",
+                            "parameters": {
+                                "Id": "1"
+                            }
+                        },
+                        {
+                            "name": "Test",
+                            "parameters": {
+                                "Id": "2"
+                            }
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "AllFilterFeature",
+                "enabled": true,
+                "conditions": {
+                    "requirement_type": "All",
+                    "client_filters": [
+                        {
+                            "name": "Test",
+                            "parameters": {
+                                "Id": "1"
+                            }
+                        },
+                        {
+                            "name": "Test",
+                            "parameters": {
+                                "Id": "2"
+                            }
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "FeatureUsesFiltersWithDuplicatedAlias",
+                "enabled": true,
+                "conditions": {
+                    "client_filters": [
+                        {
+                            "name": "DuplicatedFilterName"
+                        },
+                        {
+                            "name": "Percentage",
+                            "parameters": {
+                                "Value": 100
+                            }
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "TargetingTestFeature",
+                "enabled": true,
+                "conditions": {
+                    "client_filters": [
+                        {
+                            "name": "Targeting",
+                            "parameters": {
+                                "Audience": {
+                                    "Users": [
+                                        "Jeff",
+                                        "Alicia"
+                                    ],
+                                    "Groups": [
+                                        {
+                                            "Name": "Ring0",
+                                            "RolloutPercentage": 100
+                                        },
+                                        {
+                                            "Name": "Ring1",
+                                            "RolloutPercentage": 50
+                                        }
+                                    ],
+                                    "DefaultRolloutPercentage": 20
+                                }
+                            }
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "TargetingTestFeatureWithExclusion",
+                "enabled": true,
+                "conditions": {
+                    "client_filters": [
+                        {
+                            "name": "Targeting",
+                            "parameters": {
+                                "Audience": {
+                                    "Users": [
+                                        "Jeff",
+                                        "Alicia"
+                                    ],
+                                    "Groups": [
+                                        {
+                                            "Name": "Ring0",
+                                            "RolloutPercentage": 100
+                                        },
+                                        {
+                                            "Name": "Ring1",
+                                            "RolloutPercentage": 50
+                                        }
+                                    ],
+                                    "DefaultRolloutPercentage": 20,
+                                    "Exclusion": {
+                                        "Users": [
+                                            "Jeff"
+                                        ],
+                                        "Groups": [
+                                            "Ring0",
+                                            "Ring2"
+                                        ]
+                                    }
+                                }
+                            }
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "CustomFilterFeature",
+                "enabled": true,
+                "conditions": {
+                    "client_filters": [
+                        {
+                            "name": "CustomTargetingFilter",
+                            "parameters": {
+                                "Audience": {
+                                    "Users": [
+                                        "Jeff"
+                                    ]
+                                }
+                            }
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "VariantFeaturePercentileOn",
+                "enabled": true,
+                "variants": [
+                    {
+                        "name": "Big",
+                        "status_override": "Disabled"
+                    }
+                ],
+                "allocation": {
+                    "percentile": [
+                        {
+                            "variant": "Big",
+                            "from": 0,
+                            "to": 50
+                        }
+                    ],
+                    "seed": 1234
+                },
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeaturePercentileOff",
+                "enabled": true,
+                "variants": [
+                    {
+                        "name": "Big"
+                    }
+                ],
+                "allocation": {
+                    "percentile": [
+                        {
+                            "variant": "Big",
+                            "from": 0,
+                            "to": 50
+                        }
+                    ],
+                    "seed": 12345
+                },
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeatureAlwaysOff",
+                "enabled": false,
+                "variants": [
+                    {
+                        "name": "Big"
+                    }
+                ],
+                "allocation": {
+                    "percentile": [
+                        {
+                            "variant": "Big",
+                            "from": 0,
+                            "to": 100
+                        }
+                    ],
+                    "seed": 12345
+                },
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeatureDefaultDisabled",
+                "enabled": false,
+                "variants": [
+                    {
+                        "name": "Small",
+                        "configuration_value": "300px"
+                    }
+                ],
+                "allocation": {
+                    "default_when_disabled": "Small"
+                },
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeatureDefaultEnabled",
+                "enabled": true,
+                "variants": [
+                    {
+                        "name": "Medium",
+                        "configuration_value": {
+                            "Size": "450px",
+                            "Color": "Purple"
+                        }
+                    },
+                    {
+                        "name": "Small",
+                        "configuration_value": "300px"
+                    }
+                ],
+                "allocation": {
+                    "default_when_enabled": "Medium",
+                    "user": [
+                        {
+                            "variant": "Small",
+                            "users": [
+                                "Jeff"
+                            ]
+                        }
+                    ]
+                },
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeatureUser",
+                "enabled": true,
+                "variants": [
+                    {
+                        "name": "Small",
+                        "configuration_value": "300px"
+                    }
+                ],
+                "allocation": {
+                    "user": [
+                        {
+                            "variant": "Small",
+                            "users": [
+                                "Marsha"
+                            ]
+                        }
+                    ]
+                },
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeatureGroup",
+                "enabled": true,
+                "variants": [
+                    {
+                        "name": "Small",
+                        "configuration_value": "300px"
+                    }
+                ],
+                "allocation": {
+                    "group": [
+                        {
+                            "variant": "Small",
+                            "groups": [
+                                "Group1"
+                            ]
+                        }
+                    ]
+                },
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeatureNoVariants",
+                "enabled": true,
+                "variants": [],
+                "allocation": {
+                    "user": [
+                        {
+                            "variant": "Small",
+                            "users": [
+                                "Marsha"
+                            ]
+                        }
+                    ]
+                },
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeatureNoAllocation",
+                "enabled": true,
+                "variants": [
+                    {
+                        "name": "Small",
+                        "configuration_value": "300px"
+                    }
+                ],
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeatureAlwaysOffNoAllocation",
+                "enabled": false,
+                "variants": [
+                    {
+                        "name": "Small",
+                        "configuration_value": "300px"
+                    }
+                ],
+                "telemetry": {
+                    "enabled": true
+                }
+            },
+            {
+                "id": "VariantFeatureBothConfigurations",
+                "enabled": true,
+                "variants": [
+                    {
+                        "name": "Small",
+                        "configuration_value": "600px"
+                    }
+                ],
+                "allocation": {
+                    "default_when_enabled": "Small"
+                }
+            },
+            {
+                "id": "VariantFeatureInvalidStatusOverride",
+                "enabled": true,
+                "variants": [
+                    {
+                        "name": "Small",
+                        "configuration_value": "300px",
+                        "status_override": "InvalidValue"
+                    }
+                ],
+                "allocation": {
+                    "default_when_enabled": "Small"
+                }
+            },
+            {
+                "id": "VariantFeatureInvalidFromTo",
+                "enabled": true,
+                "variants": [
+                    {
+                        "name": "Small",
+                        "configuration_value": "300px"
+                    }
+                ],
+                "allocation": {
+                    "percentile": [
+                        {
+                            "variant": "Small",
+                            "from": "Invalid",
+                            "to": "Invalid"
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "VariantImplementationFeature",
+                "enabled": true,
+                "conditions": {
+                    "client_filters": [
+                        {
+                            "name": "Targeting",
+                            "parameters": {
+                                "Audience": {
+                                    "Users": [
+                                        "UserOmega",
+                                        "UserSigma",
+                                        "UserBeta"
+                                    ]
+                                }
+                            }
+                        }
+                    ]
+                },
+                "variants": [
+                    {
+                        "name": "AlgorithmBeta"
+                    },
+                    {
+                        "name": "Sigma",
+                        "configuration_value": "AlgorithmSigma"
+                    },
+                    {
+                        "name": "Omega"
+                    }
+                ],
+                "allocation": {
+                    "user": [
+                        {
+                            "variant": "AlgorithmBeta",
+                            "users": [
+                                "UserBeta"
+                            ]
+                        },
+                        {
+                            "variant": "Omega",
+                            "users": [
+                                "UserOmega"
+                            ]
+                        },
+                        {
+                            "variant": "Sigma",
+                            "users": [
+                                "UserSigma"
+                            ]
+                        }
+                    ]
+                }
+            },
+            {
+                "id": "OnTelemetryTestFeature",
+                "enabled": true,
+                "telemetry": {
+                    "enabled": true,
+                    "metadata": {
+                        "Tags.Tag1": "Tag1Value",
+                        "Tags.Tag2": "Tag2Value",
+                        "Etag": "EtagValue",
+                        "Label": "LabelValue"
+                    }
+                }
+            },
+            {
+                "id": "OffTelemetryTestFeature",
+                "enabled": false,
+                "telemetry": {
+                    "enabled": true
+                }
+            }
+        ]
+    }
+};
+
