+++ sdk/feature-management/tsconfig.json
@@ -1,0 +1,21 @@
+{
+    "compilerOptions": {
+        "lib": [
+            "DOM",
+            "WebWorker",
+            "ESNext"
+        ],
+        "skipDefaultLibCheck": true,
+        "module": "ESNext",
+        "moduleResolution": "Node",
+        "target": "ES2022",
+        "strictNullChecks": true,
+        "strictFunctionTypes": true,
+        "sourceMap": true,
+        "inlineSources": true
+    },
+    "exclude": [
+        "node_modules",
+        "**/node_modules/*"
+    ]
+}
