+++ sdk/feature-management/tsconfig.test.json
@@ -1,0 +1,10 @@
+{
+    "extends": "./tsconfig.json",
+    "compilerOptions": {
+        "module": "CommonJS",
+        "outDir": "./out"
+    },
+    "include": [
+        "test/*"
+    ]
+}
