--- src/common/ITargetingContext.ts
@@ -1,8 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-export interface ITargetingContext {
-    userId?: string;
-    groups?: string[];
-}
-
