--- src/index.ts
@@ -1,6 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-export { FeatureManager } from "./featureManager.js";
-export { ConfigurationMapFeatureFlagProvider, ConfigurationObjectFeatureFlagProvider, IFeatureFlagProvider } from "./featureProvider.js";
-export { IFeatureFilter } from "./filter/FeatureFilter.js";
