--- src/variant/Variant.ts
@@ -1,9 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
-export class Variant {
-    constructor(
-        public name: string,
-        public configuration: unknown
-    ) {}
-}
