--- tsconfig.test.json
@@ -1,10 +1,0 @@
-{
-    "extends": "./tsconfig.base.json",
-    "compilerOptions": {
-        "module": "CommonJS",
-        "outDir": "./out"
-    },
-    "include": [
-        "test/*"
-    ]
-}
